pyproject.toml
[project]
name = "videoglancer"
version = "0.1.1"
description = "Python port of glancer: glance through YouTube talks by stitching slides and captions."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Jorge Vicente Cantero", email = "jorgevc@fastmail.es"},
    {name = "Ruben Berenguel Montoro", email = "ruben@mostlymaths.net"},
]
dependencies = [
    "srt>=3.5",
    "Pillow>=10.0",
    "Jinja2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
]

[project.scripts]
glancer = "glancer.cli:main"
glancer-extract = "glancer.extract:main"
glancer-render = "glancer.render:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["glancer"]

[tool.pyright]
typeCheckingMode = "basic"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
